Enabled the configuration in open-api-generator-maven-plugin to use tags within the yaml spec for configuring the api groups and delegate names as well as updated the delegate implementations accordingly
